+++ tests/testthat/test-utils.R
@@ -29,13 +29,22 @@ test_that("expand_dependency", {
 })
 
 test_that("extract_deps", {
-  input <- c("ggplot2 (>= 3.1.0)", "shiny (>= 1.3.1)", "", "R (>= 3.1.0)",
-             "Biobase", "glmnet", "methods", "stats", "NA")
-  expect_identical(extract_deps(input),
-                   list(packages = list("ggplot2", "shiny", "", "R", "Biobase",
-                                        "glmnet", "methods", "stats", "NA"),
-                        versions = list("3.1.0", "1.3.1", NA_character_, "3.1.0",
-                                        NA_character_, NA_character_, NA_character_,
-                                        NA_character_, NA_character_)))
+  input <- c(
+    "ggplot2 (>= 3.1.0)", "shiny (>= 1.3.1)", "", "R (>= 3.1.0)",
+    "Biobase", "glmnet", "methods", "stats", "NA"
+  )
+  expect_identical(
+    extract_deps(input),
+    list(
+      packages = list(
+        "ggplot2", "shiny", "", "R", "Biobase",
+        "glmnet", "methods", "stats", "NA"
+      ),
+      versions = list(
+        "3.1.0", "1.3.1", NA_character_, "3.1.0",
+        NA_character_, NA_character_, NA_character_,
+        NA_character_, NA_character_
+      )
+    )
+  )
 })
-